rename(Meeting): edit exception package name
