+++ src/main/java/mvc/promiseme/meeting/exception/TransferTextException.java
@@ -1,4 +1,4 @@
-package mvc.promiseme.meeting.exceotion;
+package mvc.promiseme.meeting.exception;
 
 public class TransferTextException extends RuntimeException{
     public TransferTextException() {
